made op functional chat section
